fix collision algorithm
Fixing a bug which allowed an entity to move when standing on a blocking
tile. The problem was that the algorithm assumes that an entity
initially stands on a non-blocking tile. Failing this assumption leads
to two tile penetration values instead of one per direction tested: The
entity can move exactly one block per function call in two of the four
directions, depending on the order of the tile penetration evaluation.
The solution is to avoid calculating the penetration value against the
tile where the entity was before the movement.