change decoration item notation
Use names instead of the texture coordinates and sizes inside the level
files. This has the advantage of a more readable level file format and
that the size of a decoration tile is defined inside the source code
instead of the level file.
Additionally, Level.cpp and Editor.cpp are adapted accordingly.