rearrange sprites inside the item sprite sheet
The reason of this change are seams appearing along sprites when scaling
the game. The visual artifacts are caused by an interpolation of the
sprites with adjacent sprites inside the sprite sheet. The solution is
to introduce spacing between single sprites by rearranging them inside
the sprite sheet.
Additionally, the GUI.cpp and Editor.cpp are adapted accordingly.